Add tests/tango/conftest.py
